requirements.txt
fastapi>=0.116.0
beautifulsoup4>=4.12.0
jinja2>=3.1.4
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
mcp>=1.0.0
openai>=1.40.0
Pillow>=10.4.0
playwright>=1.47.0
pinecone>=5.4.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
uvicorn>=0.31.0
